refactor: deepen ios runner lifecycle
